requirements.txt
requests
beautifulsoup4
pandas
python-dotenv
pytest
pydantic-settings